Refactor: change Double to BigDecimal in all Records/DTOs

+++ supermarket-api/src/main/java/com/supermarket/supermarket_api/dto/SaleItemDTO.java
@@ -1,9 +1,11 @@
 package com.supermarket.supermarket_api.dto;
 
+import java.math.BigDecimal;
+
 public record SaleItemDTO(
         Long id,
         Long saleId,
         Long productId,
         Integer quantity,
-        Double subtotal
+        BigDecimal subtotal
 ) { }
